+++ packages/langchain/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @ai-sdk/langchain
 
+## 3.0.0-beta.86
+
+### Patch Changes
+
+- Updated dependencies [5a6f514]
+  - ai@7.0.0-beta.86
+
+## 3.0.0-beta.85
+
+### Patch Changes
+
+- Updated dependencies [57bf606]
+  - ai@7.0.0-beta.85
+
+## 3.0.0-beta.84
+
+### Patch Changes
+
+- Updated dependencies [90e2d8a]
+  - ai@7.0.0-beta.84
+
 ## 3.0.0-beta.83
 
 ### Patch Changes